Unsigned

+++ repository/tests/ad9910_ram.py
@@ -232,7 +232,7 @@ class AD9910RAMTest(EnvExperiment):
 
         self.n_steps = 10
 
-        self.ram_data = list(range(self.n_steps))
+        self.ram_data = np.uint32(list(range(self.n_steps)))
 
         # self.phase_start = 0.0
         # self.phase_end = 0.0
@@ -270,7 +270,7 @@ class AD9910RAMTest(EnvExperiment):
     @kernel
     def read_and_print_ram(self):
         self.core.break_realtime()
-        read_data = [np.int32(0x00)] * self.n_steps
+        read_data = [np.uint32(0x00)] * self.n_steps
         self.dds.read_ram(read_data)
 
         logger.info("RAM contents: %s", read_data)
